+++ source/ff.dxgi/source/buffer_base.h
@@ -18,7 +18,7 @@ namespace ff::dxgi
         virtual ~buffer_base() = default;
 
         virtual ff::dxgi::buffer_type type() const = 0;
-        virtual uint64_t size() const = 0;
+        virtual size_t size() const = 0;
         virtual bool writable() const = 0;
         virtual bool update(ff::dxgi::command_context_base& context, const void* data, size_t size, size_t min_buffer_size) = 0;
         virtual void* map(ff::dxgi::command_context_base& context, size_t size) = 0;
